flask-boilerplate-prod/requirements.txt
alembic==1.11.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
bombaysoftwares-pysupp==1.0.4
boto3==1.24.70
botocore==1.27.70
cfgv==3.3.1
click==8.1.3
crontab==1.0.1
Deprecated==1.2.13
distlib==0.3.6
exceptiongroup==1.1.1
filelock==3.12.0
Flask==2.3.2
Flask-Limiter==3.3.0
Flask-Mail==0.9.1
Flask-Migrate==2.6.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.5.1
flask-swagger-ui==4.11.1
freezegun==1.2.2
greenlet==2.0.2
hashids==1.3.1
identify==2.5.24
importlib-metadata==6.8.0
importlib-resources==5.12.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
limits==3.5.0
Mako==1.2.4
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
nodeenv==1.8.0
numpy==1.24.4
ordered-set==4.1.0
packaging==23.1
pkgutil_resolve_name==1.3.10
platformdirs==3.5.1
pluggy==1.0.0
pre-commit==3.3.2
psycopg2-binary==2.9.6
Pygments==2.15.1
PyJWT==2.7.0
pytest==7.3.1
pytest-ordering==0.6
python-dateutil==2.8.2
python-magic==0.4.27
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0
redis==4.5.5
referencing==0.30.2
rich==13.3.5
rpds-py==0.10.3
rq==1.14.1
rq-scheduler==0.13.1
s3transfer==0.6.2
six==1.16.0
SQLAlchemy==1.4.40
svgwrite==1.4.3
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.16
virtualenv==20.23.0
Werkzeug==2.3.4
wrapt==1.15.0
zipp==3.16.2
